+++ .gitignore
@@ -1,4 +1,5 @@
 # Hospital Project
+.idea
 Hospital/.idea/
 Hospital/out/
 Hospital/target/
