--- Hospital/src/main/java/Authorizer.java
@@ -1,7 +1,0 @@
-package main.java;
-
-import main.java.User;
-
-public class Authorizer {
-    private User[] users;
-}
