+++ Hospital/src/main/java/controllers/Menu.java
@@ -1,0 +1,4 @@
+package main.java.controllers;
+
+public class Menu {
+}
