+++ Hospital/src/main/java/controllers/resource_controllers/Authorizer.java
@@ -1,0 +1,7 @@
+package main.java.controllers.resource_controllers;
+
+import main.java.users.User;
+
+public class Authorizer {
+    private User[] users;
+}
